+++ src/server/game/Spells/Auras/SpellAuraEffects.cpp
@@ -1653,7 +1653,7 @@ int32 AuraEffect::CalculateAmount(Unit* caster, int32 &m_aura_amount)
             if(m_spellInfo->AttributesEx10 & SPELL_ATTR10_STACK_DAMAGE_OR_HEAL && GetBase()->GetMaxDuration() != -1 && caster)
             {
                 int32 tempAmount = GetTotalTicks() * (amount + m_aura_amount);
-                if(tempAmount <= caster->GetMaxHealth())
+                if(tempAmount <= (int32)caster->GetMaxHealth())
                 {
                     amount += m_aura_amount;
                     m_crit_amount = amount * 2;
@@ -2353,7 +2353,7 @@ void AuraEffect::CleanupTriggeredSpells(Unit* target)
             return;
 
         if (tProto->StackAmount > 1)
-            if (GetTotalTicks() < tProto->StackAmount)
+            if (GetTotalTicks() < (int32)tProto->StackAmount)
                 return;
     }
