+++ src/server/game/Spells/SpellEffects.cpp
@@ -1598,7 +1598,7 @@ void Spell::EffectDummy(SpellEffIndex effIndex)
         {
             int32 countDamage = int32(m_caster->GetDamageCounterInPastSecs(5, DAMAGE_TAKEN_COUNTER) * 0.20f);
 
-            if (m_caster->CountPctFromMaxHealth(7) > countDamage)
+            if ((int32)m_caster->CountPctFromMaxHealth(7) > countDamage)
                 bp = m_caster->CountPctFromMaxHealth(7);
             else
                 bp = countDamage;
